mcp_server Phase 1: extract config.py, client.py, models.py from the monolith
Splits the 1062-line tools/mcp_server/server.py per prog.txt Sezione 3's
plan: config.py (env-derived settings, tool-annotation presets), client.py
(shared httpx.AsyncClient, _request, and a new @catch_errors decorator
replacing ~20 repeated try/except-return-_handle_error(e) blocks), models.py
(all 16 Pydantic input schemas). server.py keeps the MCPServer instance,
the 21 tool functions, and the handful of helpers only they use (image
saving/pruning, statevector/probability truncation, molecule-name
resolution).

KERNEL_URL and _TEST_TRANSPORT live in client.py rather than config.py --
tests/integration/test_mcp_server.py mutates both directly to exercise the
unreachable-kernel and in-process-ASGI-transport paths, and that mutation
only works if the value's single source of truth lives in the same module
that reads it (_get_client/_request). Updated the test suite's own
mcp_adapter.KERNEL_URL/_TEST_TRANSPORT/_shared_client/_get_client
references to mcp_client.* accordingly -- verified this catches a real
break (a stale monkeypatch.setattr(mcp_adapter, "_get_client", ...) would
raise AttributeError once _get_client no longer lives in server.py).

Verified: all 39 tests in test_mcp_server.py pass: real DenseSVSimulator/
PennyLane execution via the in-process ASGI transport, not mocked physics.
No behavior change for callers -- dynamic per-endpoint timeouts and the
non-JSON-error-response fallback already existed pre-refactor and are
moved as-is, not rebuilt (see prog.txt Sezione 3.1, corrected 2026-08-20
after checking the diagnosis against current code).

Phase 2 (utils/images.py, utils/truncation.py, utils/cache.py with TTL)
and Phase 3 (splitting the 21 tools into tools/*.py by topic) are separate
follow-up PRs, per prog.txt's "one PR per phase" rule.
